Move BodyFaceGestureAnalysis to own file
